Rewrite `PathExpression` visitor code
Instead of having a single visitor for `PathExpression` it is more useful to know if it is the path of a simple mustache, a block, a modifier, etc.

Depending on the parent element we can assume certain things, e.g. inside an `AttrNode` the mustache path can't be a component invocation.

This rewrite also fixes the wrong priority order where properties had priority over components and helpers, which is different in an actual app. If a component/helper with a corresponding name is found than the prefixing will be skipped. This removes the reliance on `getTelemetryFor()` which helps bring us closer to "pods" support.

Due to the different implementation this also removes the `dontAssumeThis` option from the codemod.